Refactor a GA4 account create info notice to use the new design.
