+++ assets/js/modules/analytics-4/components/common/index.js
@@ -24,6 +24,6 @@ export { default as WebDataStreamSelect } from './WebDataStreamSelect';
 export { default as PropertySelect } from './PropertySelect';
 export { default as TrackingExclusionSwitches } from './TrackingExclusionSwitches';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
-export { default as Notice } from './Notice';
+export { default as Notice } from './CreatePropertyInfoNotice';
 export { default as WebDataStreamNameInput } from './WebDataStreamNameInput';
 export { default as ZeroDataMessage } from './ZeroDataMessage';
